Change path prefix for butler server
For the purposes of the RSP, Butler is considered part of the "API
Aspect", so the path to it needs to start with /api/.  Handlers were
re-organized to be grouped under an APIRouter, since this prefix will
need to be configurable in the future.
